pyproject.toml
[project]
name = "lightninglinks"
version = "0.2.1"
description = "An automatic obsidian note linker and smart assistant"
authors = [{ name = "EthanDGee", email = "workethangee@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
  "openai (>=2.15.0,<3.0.0)",
  "transformers (>=4.57.3,<5.0.0)",
  "numpy (>=2.4.1,<3.0.0)",
  "pydantic (>=2.12.5,<3.0.0)",
  "torch (>=2.9.1,<3.0.0)",
  "sentence-transformers (>=5.2.0,<6.0.0)",
  "python-dotenv (>=1.2.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
